Add NumPy v2 compatible sign function (#1984)
Add new `np.sign` implementations to match the new definition in NumPy
v2.0. Fixes #1915

Remove the numpy folder from stdlib. In C/Fortran there is no need to
link sign with NumPy.

**Commit Summary**
- Add `np.sign` implementations matching NumPy v2
- Choose the correct `np.sign` implementation depending on the NumPy
version
- Correct the imports used by `isinf`, `isnan` and `isfinite`
- Remove dead code in c printer
