+++ pyccel/stdlib/math/pyc_math_c.c
@@ -4,7 +4,6 @@
 /* -------------------------------------------------------------------------------------- */
 
 #include "pyc_math_c.h"
-#include <math.h>
 
 /*---------------------------------------------------------------------------*/
 int64_t                     pyc_factorial(int64_t n)
@@ -43,3 +42,27 @@ extern inline double        pyc_degrees(double radians);
 extern inline int64_t       pyc_modulo(int64_t a, int64_t b);
 /*---------------------------------------------------------------------------*/
 extern inline double        pyc_fmodulo(double a, double b);
+
+/* numpy.sign for float, double and integers */
+long long int isign(long long int x)
+{
+    return (x>0) - (x<0);
+}
+
+/* numpy.sign for float, double and integers */
+double  fsign(double x)
+{
+    return (double)((x>0) - (x<0));
+}
+
+/* numpy.sign for complex for NumPy v1 */
+double complex csgn(double complex x)
+{
+    return x ? ((!creal(x) && cimag(x) < 0) || (creal(x) < 0) ? -1 : 1) : 0;
+}
+
+double complex csign(double complex x)
+{
+    double absolute = cabs(x);
+    return ((absolute == 0) ? 0.0 : (x / absolute));
+}
